+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Added functionality to initialize metrics at 0, so instrumented functions can be registered.
 - Fixed an issue where decorators were changing `this` values for the methods they'd be wrapping, breaking them.
 - Improved how `getModulePath` utility works, passing stack trace as structured data, and making it more robust.