+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added functionality to initialize metrics at 0, so instrumented functions can be registered.
 - Fixed an issue where decorators were changing `this` values for the methods they'd be wrapping, breaking them.
 - Improved how `getModulePath` utility works, passing stack trace as structured data, and making it more robust.
+- Change duration metric to report seconds instead of milliseconds
 
 ## [v0.5.3] - 2023-06-26