+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed an issue where decorators were changing `this` values for the methods they'd be wrapping, breaking them.
 - Improved how `getModulePath` utility works, passing stack trace as structured data, and making it more robust.
 - Change duration metric to report seconds instead of milliseconds
+- Push metrics to gateway "eagerly" when pushInterval is set to 0
 
 ## [v0.5.3] - 2023-06-26