+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added functionality to initialize metrics at 0, so instrumented functions can be registered.
 - Fixed an issue where decorators were changing `this` values for the methods they'd be wrapping, breaking them.
 - Improved how `getModulePath` utility works, passing stack trace as structured data, and making it more robust.
+- Measure function duration in seconds, and update the histogram buckets in prometheus accordingly
 
 ## [v0.5.3] - 2023-06-26
 
@@ -14,7 +15,7 @@
 
 ## [v0.5.2] - 2023-06-16
 
-### Added 
+### Added
 
 - added `docsOutputFormat` config parameter to @autometrics/typescript-plugin. This is for communicating autometrics information from typescript to the autometrics extension.
 
@@ -39,6 +40,7 @@
 - added links to Grafana dashboards in the README
 
 ### Changed
+
 - `launch.json` configurations for easier TypeScript plugin debugging workflow
 
 ## [v0.3] - 2023-04-23